backend/requirements.txt
django
djangorestframework
pandas
matplotlib
django-cors-headers
djangorestframework-simplejwt
